Moved all services to its own service-directory
