+++ lib/create-app.js
@@ -12,8 +12,8 @@ const authProxy = require('./handlers/auth-proxy')
 const SolidHost = require('./models/solid-host')
 const AccountManager = require('./models/account-manager')
 const vhost = require('vhost')
-const EmailService = require('./models/email-service')
-const TokenService = require('./models/token-service')
+const EmailService = require('./services/email-service')
+const TokenService = require('./services/token-service')
 const capabilityDiscovery = require('./capability-discovery')
 const API = require('./api')
 const errorPages = require('./handlers/error-pages')
