+++ lib/requests/create-account-request.js
@@ -3,7 +3,7 @@
 const AuthRequest = require('./auth-request')
 const WebIdTlsCertificate = require('../models/webid-tls-certificate')
 const debug = require('../debug').accounts
-const blacklistService = require('../common/blacklist-service')
+const blacklistService = require('../services/blacklist-service')
 
 /**
  * Represents a 'create new user account' http request (either a POST to the
