+++ lib/services/email-service.js
@@ -2,7 +2,7 @@
 
 const nodemailer = require('nodemailer')
 const path = require('path')
-const debug = require('./../debug').email
+const debug = require('../debug').email
 
 /**
  * Models a Nodemailer-based email sending service.
